third_party/WebKit/LayoutTests/fast/dom/shadow/css-focus-pseudo-match-shadow-host.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
<style>
div {
    background-color: white;
}

div#shadow-host:focus {
    background-color: green;
}
</style>
<script>
function backgroundColorOf(selector) {
    return window.getComputedStyle(getNodeInTreeOfTrees(selector)).backgroundColor;
}

function backgroundColorShouldBe(selector, expected) {
    shouldBeEqualToString('backgroundColorOf(\'' + selector + '\')', expected);
}

function testSingleShadow() {
    debug('testing single shadow DOM tree');

    // Setting up the DOM tree
    var sandbox = document.querySelector('#sandbox');
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {},
            createDOM('input', {'id': 'outer-input'}),
            createDOM('div', {'id': 'shadow-host'},
                createShadowRoot(
                    createDOM('div', {'id': 'inner-div'},
                        document.createTextNode('Blink')),
                    createDOM('input', {'id': 'inner-input'})))));
    sandbox.offsetTop;

    var host = getNodeInTreeOfTrees('shadow-host');
    var input = getNodeInTreeOfTrees('shadow-host/inner-input');
    var outerInput = document.querySelector('#outer-input');

    // First check the original behavior.
    // By default, shadow host should not reflect the focus status inside shadow tree.
    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    host.tabIndex = 0;
    host.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    // The following 4 cases cover tabIndex(0 or none)/tabStop(true or false) combination.
    // Only when tabindex>=0 and tabStop=false should change the appearance.
    host.tabIndex = 0;
    host.tabStop = false;
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');

    host.tabStop = true;
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    host.removeAttribute('tabindex');
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    host.tabStop = false;
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    // Focus outside the shadow tree should not affect.
    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
}

function testShadowTreeOfTrees() {
    debug('testing shadow tree of trees');

    // Setting up the DOM tree
    var sandbox = document.querySelector('#sandbox');
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {},
            createDOM('input', {'id': 'outer-input'}),
            createDOM('div', {'id': 'shadow-host'},
                createShadowRoot(
                    createDOM('style', {},
                        document.createTextNode('div { background-color: yellow; } div#inner-shadow-host:focus { background-color: blue; }')),
                    createDOM('div', {'id': 'inner-div'},
                        document.createTextNode('Blink')),
                    createDOM('input', {'id': 'inner-input'}),
                    createDOM('div', {'id': 'inner-shadow-host'},
                        createShadowRoot(
                            createDOM('div', {'id': 'inner-div2'},
                                document.createTextNode('Blink')),
                            createDOM('input', {'id': 'inner-input2'})))))));
    sandbox.offsetTop;

    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(255, 255, 0)');

    var host = getNodeInTreeOfTrees('shadow-host');
    var innerHost = getNodeInTreeOfTrees('shadow-host/inner-shadow-host');
    var input = getNodeInTreeOfTrees('shadow-host/inner-input');
    var input2 = getNodeInTreeOfTrees('shadow-host/inner-shadow-host/inner-input2');
    var outerInput = document.querySelector('#outer-input');

    // The default behavior.
    input2.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(255, 255, 0)');


    // Only first-level shadow host should be affected.
    host.tabIndex = 0;
    host.tabStop = false;

    input2.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(255, 255, 0)');

    // Both first-level and second-level shadow hosts should be affected.
    innerHost.tabIndex = 0;
    innerHost.tabStop = false;
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(0, 0, 255)');

    // The :focus should propagate only upwards, not downwards.
    input.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(255, 255, 0)');

    // Focus outside the shadow tree should not affect.
    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    backgroundColorShouldBe('shadow-host/inner-shadow-host', 'rgb(255, 255, 0)');
}

function testDistributedNodes() {
    debug('testing distributed nodes');

    // Setting up the DOM tree
    var sandbox = document.querySelector('#sandbox');
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {},
            createDOM('input', {'id': 'outer-input'}),
            createDOM('div', {'id': 'shadow-host'},
                createDOM('input', {'id': 'light-input'}),
                createShadowRoot(
                    createDOM('content', {}),
                    createDOM('div', {'id': 'inner-div'},
                        document.createTextNode('Blink')),
                    createDOM('input', {'id': 'older-input'})),
                createShadowRoot(
                    createDOM('shadow', {}),
                    createDOM('div', {'id': 'inner-div2'},
                        document.createTextNode('Blink')),
                    createDOM('input', {'id': 'younger-input'})))));
    sandbox.offsetTop;

    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    var host = getNodeInTreeOfTrees('shadow-host');
    var outerInput = getNodeInTreeOfTrees('outer-input');
    var lightInput = getNodeInTreeOfTrees('light-input');
    var olderInput = getNodeInTreeOfTrees('shadow-host/older-input');
    var youngerInput = getNodeInTreeOfTrees('shadow-host//younger-input');

    lightInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    host.tabIndex = 0;
    host.tabStop = false;

    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    lightInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');

    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    olderInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');

    outerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(255, 255, 255)');
    youngerInput.focus();
    backgroundColorShouldBe('shadow-host', 'rgb(0, 128, 0)');
}

function init() {
    if (!window.eventSender)
        testFailed('No eventSender');

    testSingleShadow();
    testShadowTreeOfTrees();
    testDistributedNodes();
}

window.onload = init;
</script>
<body>
<div id="sandbox"></div>
</body>
